Added Favorite Functionality
Added a button to the detail layout that the user
can click to add or remove a movie from the local
database. When clicked, the button will change to
display the appropriate text. In order to initialize
the correct text into the button, the main activity
will determine if the movie is a favorite and send
that information through the intent. The actual
determination is another task left to do.