Tested and Fixed some bugs
Upon testing, found the following bugs. MovieDetailActivity
would not load trailers or reviews (the loaders were only
being init, not actually loaded: fixed). Favorite button
worked, but only once (the Boolean variable isFavorite was
not being updated with each click: fixed). MovieDetailActivity
is not maintaining life cycle independence when rotated into
landscape mode.

One improvement to MainActivity will be to have the live data
update the recycler view when the data changes, but only when
the spinner is set to the appropriate setting. Also whenever
the recycler view is filled with the favorites, the loading
indicator should always be hidden.