Tested and Fixed some bugs
MovieDetailActivity was not maintaining life cycle
independence when rotated into landscape mode. The problem
was using an additional persistableBundle in
onSavedInstantState. After removing, life cycle independence
worked fine.

Now, in MainActivity, the live data updates the recycler
view when the data changes, but only when the spinner is
set to the appropriate setting. Also whenever the recycler
view is filled with the favorites, the loading indicator
is hidden.